requirements.txt
django>=4.2,<5.0
gunicorn>=21.2,<22.0
dj-database-url>=2.0,<3.0
whitenoise[brotli]>=6.0,<7.0
psycopg; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"
pandas==2.1.1
requests==2.31.0
django-cors-headers==4.2.0
djangorestframework==3.14.0